requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==3.2.15
django-filter==22.1
django-social-share==2.3.0
djangorestframework==3.14.0
idna==3.4
importlib-metadata==5.0.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.9.0
